fastapi-projects/14-vakeel-contracts/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
httpx
PyPDF2
python-multipart
